APIv4 - use entityTypes event to load custom records
This standardizes things to use the ActionObjectProvider service more like it was intended,
with getEntityNames now returning what it's supposed to, and removing a direct-cache-access hack
from CoreUtil, in favor of using the service.
